check explain also updates explain in metrics
